# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Webstore Customer API",
        "description": "Dokumentasi endpoint API customer untuk autentikasi dan fitur terkait dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Default API server"
        }
    ],
    "paths": {
...
        "/api/dashboard/orders/{order}/payment-status": {
            "post": {
                "tags": [
                    "Dashboard Orders"
                ],
                "summary": "Sinkronisasi status pembayaran order",
                "operationId": "db6cf974a516a3babaa28954926a88be",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status pembayaran berhasil diperiksa",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Status pembayaran sudah diperbarui.",
                                    "data": {
                                        "order": {
                                            "id": 1201,
                                            "order_no": "ORD-20260301-ABC123",
                                            "status": "paid"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tindakan tidak diizinkan."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Data tidak valid.",
                                    "errors": {
                                        "field": [
                                            "Field wajib diisi."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/orders/{order}/pay-now": {
            "post": {
                "tags": [
                    "Dashboard Orders"
                ],
                "summary": "Generate token Midtrans pay now untuk order",
                "operationId": "bb223687ef2596a8d7885e9b24fb384e",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token pay now berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Token pembayaran Midtrans berhasil dibuat.",
                                    "snapToken": "midtrans-snap-token",
                                    "orderId": 1201,
                                    "orderNo": "ORD-20260301-ABC123"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tindakan tidak diizinkan."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Data tidak valid.",
                                    "errors": {
                                        "field": [
                                            "Field wajib diisi."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/orders/{order}/invoice": {
            "get": {
                "tags": [
                    "Dashboard Orders"
                ],
                "summary": "Download invoice order customer",
                "description": "Menghasilkan file PDF invoice untuk order customer yang sudah dibayar.",
                "operationId": "fecbdb46d996b340a1bfbef9cadad2b4",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "preview",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice PDF berhasil dibuat",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary",
                                    "example": "%PDF-1.7 ...binary invoice content..."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tindakan tidak diizinkan."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice tidak tersedia",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Invoice hanya tersedia untuk pesanan yang sudah dibayar."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}